refactor: standardize Utils header inclusion and implement utility functions
