+++ src/handlers/ErrorResponseBuilder.cpp
@@ -1,5 +1,5 @@
 #include "ErrorResponseBuilder.hpp"
-#include "utils.hpp"
+#include "Utils.hpp"
 
 #include <sstream>
 #include <string>
