+++ src/handlers/RedirectHandler.cpp
@@ -1,6 +1,6 @@
 #include "RedirectHandler.hpp"
 #include "ErrorResponseHandler.hpp"
-#include "utils.hpp"
+#include "Utils.hpp"
 
 #include <string>
