+++ src/http/RequestInspector.cpp
@@ -1,7 +1,7 @@
 #include "RequestInspector.hpp"
 #include "ChunkedDecoder.hpp"
 #include "HttpMessageUtils.hpp"
-#include "utils.hpp"
+#include "Utils.hpp"
 
 #include <string>
 #include <sstream>
