+++ src/utils/Utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.hpp"
+#include "Utils.hpp"
 
 #include <cctype>
 #include <sstream>
